Allow resolveTransportVersionConflict to run without compilation (#140851)
Transport version generation relies on finding uses of
`TransportVersion.fromName` to identify the definition name to generate.
In the case of resolving a merge conflict, a definition has already been
generated, and the merge base is used to find differences in definition
files. So in the resolve conflict case, we don't actually need to find
uses of `TransportVersion.fromName`, and thus don't need compilation.

This commit moves most of the logic from the generate task into a base
class, which is then specialized by the normal generation task and the
resolve conflict task. By avoiding compilation running
resolveTransportVersionConflict should be much faster, and can run even
if there are other merge conflicts which may break compilation.

relates ES-14006
